refact mutebutton done
